requirements.txt
pyserial
paho-mqtt
PyYAML
